Split AST source into multiple compile units
