+++ Makefile.conf
@@ -2,6 +2,13 @@
 
 SOURCES = \
 	ast.cpp \
+	ast/nodes.cpp \
+	ast/blocks.cpp \
+	ast/values.cpp \
+	ast/selectors.cpp \
+	ast/containers.cpp \
+	ast/statements.cpp \
+	ast/expressions.cpp \
 	base64vlq.cpp \
 	bind.cpp \
 	color_maps.cpp \
