+++ configure.ac
@@ -25,6 +25,7 @@ AC_GNU_SOURCE
 # Check fails on Travis, but it works fine
 # AX_CXX_COMPILE_STDCXX_11([ext],[optional])
 AC_CHECK_TOOL([AR], [ar], [false])
+AS_CASE([$target], [*-*-mingw32], [is_mingw32=yes], [is_mingw32=no])
 if test "x$is_mingw32" != "xyes"; then
   AC_CHECK_TOOL([DLLTOOL], [dlltool], [false])
   AC_CHECK_TOOL([DLLWRAP], [dllwrap], [false])
@@ -46,7 +47,6 @@ AC_CHECK_FUNCS([floor getcwd strtol])
 AC_ARG_ENABLE(tests, AS_HELP_STRING([--enable-tests], [enable testing the build]),
               [enable_tests="$enableval"], [enable_tests=no])
 
-AS_CASE([$target], [*-*-mingw32], [is_mingw32=yes], [is_mingw32=no])
 AM_CONDITIONAL(COMPILER_IS_MINGW32, test "x$is_mingw32" = "xyes")
 
 dnl The dlopen() function is in the C library for *BSD and in
