+++ src/ast/common.hpp
@@ -1,0 +1,56 @@
+#ifndef SASS_AST_COMMON_H
+#define SASS_AST_COMMON_H
+
+#include "sass.h"
+#include "sass/context.h"
+
+namespace Sass {
+
+  // easier to search with name
+  const bool DELAYED = true;
+
+  // ToDo: should this really be hardcoded
+  // Note: most methods follow precision option
+  const double NUMBER_EPSILON = 0.00000000000001;
+
+  // ToDo: where does this fit best?
+  // We don't share this with C-API?
+  class Operand {
+    public:
+      Operand(Sass_OP operand, bool ws_before = false, bool ws_after = false)
+      : operand(operand), ws_before(ws_before), ws_after(ws_after)
+      { }
+    public:
+      enum Sass_OP operand;
+      bool ws_before;
+      bool ws_after;
+  };
+
+
+  //////////////////////////////////////////////////////////////////////
+  // `hash_combine` comes from boost (functional/hash):
+  //////////////////////////////////////////////////////////////////////
+  // http://www.boost.org/doc/libs/1_35_0/doc/html/hash/combine.html
+  // Boost Software License - Version 1.0
+  // http://www.boost.org/users/license.html
+  //////////////////////////////////////////////////////////////////////
+  template <typename T>
+  void hash_combine (std::size_t& seed, const T& val)
+  {
+    seed ^= std::hash<T>()(val) + 0x9e3779b9
+             + (seed<<6) + (seed>>2);
+  }
+  //////////////////////////////////////////////////////////////////////
+
+
+  inline bool is_ns_eq(const std::string& l, const std::string& r)
+  {
+    if (l.empty() && r.empty()) return true;
+    else if (l.empty() && r == "*") return true;
+    else if (r.empty() && l == "*") return true;
+    else return l == r;
+  }
+
+}
+
+#endif
