+++ src/ast/containers.cpp
@@ -1,0 +1,28 @@
+#include "../sass.hpp"
+#include "containers.hpp"
+#include "../ast.hpp"
+#include "../context.hpp"
+#include "../node.hpp"
+#include "../extend.hpp"
+#include "../emitter.hpp"
+#include "../color_maps.hpp"
+#include <set>
+#include <iomanip>
+#include <iostream>
+#include <algorithm>
+#include <functional>
+#include <cctype>
+#include <locale>
+
+namespace Sass {
+
+  static Null sass_null(Sass::Null(ParserState("null")));
+
+  Expression* Hashed::at(Expression* k) const
+  {
+    if (elements_.count(k))
+    { return elements_.at(k); }
+    else { return &sass_null; }
+  }
+
+}
